requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.4
fonttools==4.42.1
fsspec==2023.9.0
huggingface-hub==0.17.1
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
